pyproject.toml
[tool.poetry]
name = "self-learning-platform"
version = "0.1.0"
description = ""
authors = ["Pinchuk Sergey <Pinboyer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
django-stubs = "^5.1.1"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"
django-filter = "^24.3"
unidecode = "^1.3.8"
python-dotenv = "^1.0.1"
black = "^24.10.0"
coverage = "^7.6.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
